EleganceAI_V1-Dgango/requirements.txt
beautifulsoup4==4.12.3
Django==5.1.1
google-ai-generativelanguage==0.6.6
google-generativeai==0.7.2
huggingface-hub==0.24.7
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-google-genai==2.0.0
langchain-huggingface==0.1.0
langchain-text-splitters==0.3.0
numpy==1.26.4
python-dotenv==1.0.1
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.0
tokenizers==0.19.1
transformers==4.44.2

